third_party/blink/web_tests/fast/forms/time/time-setrangetext.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../resources/common-setrangetext.js"></script>
</head>
<body>
<script>
description("Test setRangeText() method is not available in time inputs.");
runTestsShouldFail("input", { type: "time" });
</script>
</body>
</html>
